feat(SHS-6637): Add a 'reset' button to SiteImprove details page
