+++ docroot/modules/humsci/hs_siteimprove/hs_siteimprove.routing.yml
@@ -11,7 +11,7 @@ hs_siteimprove.api_settings_form:
 hs_siteimprove.site_details:
   path: '/admin/config/services/siteimprove/details'
   defaults:
-    _controller: '\Drupal\hs_siteimprove\Controller\SiteImproveController::siteDetails'
+    _form: '\Drupal\hs_siteimprove\Form\SiteImproveDetailsForm'
     _title: 'SiteImprove Site Details'
   requirements:
     _permission: 'administer siteimprove'
